requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.3.0
pillow==10.4.0
torch>=2.1.0,<3.0.0
torchvision>=0.16.0,<1.0.0
numpy>=1.25.0,<3.0.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
python-multipart==0.0.20
